fixed major bug with login and now correctly shows users are logged in or out
